requirements/base.txt
asgiref==3.8.1
Django==5.1.1
django-authtools==2.0.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
tzdata==2024.1